backend/requirements.txt
livekit-agents
python-dotenv
livekit-api
flask[async]
flask
faiss-cpu
sentence-transformers
livekit-plugins-silero
numpy==
flask-cors
uvicorn